+++ sydtest/src/Test/Syd/Def/TestDefM.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DeriveTraversable #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
@@ -14,14 +15,16 @@
 module Test.Syd.Def.TestDefM where
 
 import Control.Monad
-import Control.Monad.RWS.Strict
 import Control.Monad.Random
+import Control.Monad.Reader
+import Control.Monad.Writer.Strict
 import Data.DList (DList)
 import qualified Data.DList as DList
 import Data.Kind
 import Data.Maybe
 import Data.Text (Text)
 import qualified Data.Text as T
+import GHC.Generics (Generic)
 import System.Random.Shuffle
 import Test.QuickCheck.IO ()
 import Test.Syd.OptParse
@@ -50,9 +53,22 @@ type TestDef outers inner = TestDefM outers inner ()
 --
 -- In practice, all of these three parameters should be '()' at the top level.
 newtype TestDefM (outers :: [Type]) inner result = TestDefM
-  { unTestDefM :: RWST TestRunSettings (TestForest outers inner) () IO result
+  { unTestDefM :: WriterT (TestForest outers inner) (ReaderT TestDefEnv IO) result
   }
-  deriving (Functor, Applicative, Monad, MonadIO, MonadReader TestRunSettings, MonadWriter (TestForest outers inner), MonadState ())
+  deriving
+    ( Functor,
+      Applicative,
+      Monad,
+      MonadIO,
+      MonadReader TestDefEnv,
+      MonadWriter (TestForest outers inner)
+    )
+
+data TestDefEnv = TestDefEnv
+  { testDefEnvDescriptionPath :: ![Text],
+    testDefEnvTestRunSettings :: !TestRunSettings
+  }
+  deriving (Show, Eq, Generic)
 
 execTestDefM :: Settings -> TestDefM outers inner result -> IO (TestForest outers inner)
 execTestDefM sets = fmap snd . runTestDefM sets
@@ -60,7 +76,12 @@ execTestDefM sets = fmap snd . runTestDefM sets
 runTestDefM :: Settings -> TestDefM outers inner result -> IO (result, TestForest outers inner)
 runTestDefM sets defFunc = do
   let func = unTestDefM defFunc
-  (a, _, testForest) <- runRWST func (toTestRunSettings sets) ()
+  let testDefEnv =
+        TestDefEnv
+          { testDefEnvDescriptionPath = [],
+            testDefEnvTestRunSettings = toTestRunSettings sets
+          }
+  (a, testForest) <- runReaderT (runWriterT func) testDefEnv
   let testForest' = filterTestForest (settingFilter sets) testForest
   stdgen <- case settingSeed sets of
     FixedSeed seed -> pure $ mkStdGen seed
@@ -70,6 +91,21 @@ runTestDefM sets defFunc = do
           then evalRand (randomiseTestForest testForest') stdgen
           else testForest'
   pure (a, testForest'')
+
+-- | Get the path of 'describe' strings upwards.
+--
+-- Note that using this function makes tests less movable, depending on what
+-- you do with these strings.
+-- For example, if you use these strings to define the path to a golden test
+-- file, then that path will change if you move the tests somewhere else.
+-- This combines unfortunately with the way @sydtest-discover@ makes the module
+-- name part of this path.
+-- Indeed: moving your tests to another module will change their path as well,
+-- if you use @sydtest-discover@.
+-- Also note that while test forests can be randomised, their description path
+-- upwards will not, because of how trees are structured.
+getTestDescriptionPath :: TestDefM outers inner [Text]
+getTestDescriptionPath = asks testDefEnvDescriptionPath
 
 toTestRunSettings :: Settings -> TestRunSettings
 toTestRunSettings Settings {..} =